refactor: 상세 조회, 게스트 조회, 참여 정보 조회 등의 책임을 Mapper로 분리
