+++ src/main/java/umc/cockple/demo/domain/exercise/converter/ExerciseLifecycleMapper.java
@@ -2,8 +2,11 @@ package umc.cockple.demo.domain.exercise.converter;
 
 import org.springframework.stereotype.Component;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
+import umc.cockple.demo.domain.exercise.domain.ExerciseAddr;
 import umc.cockple.demo.domain.exercise.dto.ExerciseCreateDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseDeleteDTO;
+import umc.cockple.demo.domain.exercise.dto.ExerciseDetailDTO;
+import umc.cockple.demo.domain.exercise.dto.ExerciseEditDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseUpdateDTO;
 
 @Component
@@ -72,6 +75,38 @@ public class ExerciseLifecycleMapper {
         return ExerciseUpdateDTO.Response.builder()
                 .exerciseId(exercise.getId())
                 .updatedAt(exercise.getUpdatedAt())
+                .build();
+    }
+
+    public ExerciseDetailDTO.Response toDetailResponse(
+            boolean isManager,
+            ExerciseDetailDTO.ExerciseInfo exerciseInfo,
+            ExerciseDetailDTO.ParticipantGroup participantGroup,
+            ExerciseDetailDTO.WaitingGroup waitingGroup) {
+
+        return ExerciseDetailDTO.Response.builder()
+                .isManager(isManager)
+                .info(exerciseInfo)
+                .participants(participantGroup)
+                .waiting(waitingGroup)
+                .build();
+    }
+
+    public ExerciseEditDetailDTO.Response toEditDetailResponse(Exercise exercise) {
+        ExerciseAddr addr = exercise.getExerciseAddr();
+
+        return ExerciseEditDetailDTO.Response.builder()
+                .date(exercise.getDate())
+                .buildingName(addr.getBuildingName())
+                .roadAddress(addr.getStreetAddr())
+                .latitude(addr.getLatitude())
+                .longitude(addr.getLongitude())
+                .startTime(exercise.getStartTime())
+                .endTime(exercise.getEndTime())
+                .maxCapacity(exercise.getMaxCapacity())
+                .allowMemberGuestsInvitation(exercise.getPartyGuestAccept())
+                .allowExternalGuests(exercise.getOutsideGuestAccept())
+                .notice(exercise.getNotice())
                 .build();
     }
 }
